fix warnings
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@51245 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ input3D/impl/src/main/java/org/geogebra/desktop/geogebra3D/input3D/intelRealSense/Socket.java
@@ -663,7 +663,9 @@ public class Socket {
 					Log.debug("Successful update");
 					showMessage(loc.getMenu("RealSense.UpdatedRuntimes"),
 							loc.getMenu("RealSenseUpdated2"));
-					destWebSetup.delete();
+					if (destWebSetup != null) {
+						destWebSetup.delete();
+					}
 				}
 
 				installRuntimes = false;
